+++ src/crate.ts
@@ -1,8 +1,6 @@
 import _debug from 'debug'
 import _ from 'lodash/fp.js'
-import ms from 'ms'
 import fetch from 'node-fetch'
-import retry, { type Options } from 'p-retry'
 
 import {
   getAuthHeader,
@@ -50,18 +48,6 @@ export interface QueryOptions {
 
 const defaultConfig = { sqlEndpoint: 'http://localhost:4200/_sql' }
 
-const retryOptions: Options = {
-  minTimeout: ms('30s'),
-  maxTimeout: ms('1h'),
-  // This translates to retrying for up to 24 hours. It takes 7 retries
-  // to reach the maxTimeout of 1 hour.
-  retries: 30,
-  shouldRetry(error) {
-    // Don't retry if we get this exception
-    return !error.message.includes('DuplicateKeyException')
-  },
-}
-
 export const crate = (config?: CrateConfig) => {
   const { sqlEndpoint, auth } = _.defaults(defaultConfig, config)
   const authHeader = auth && getAuthHeader(auth)
@@ -70,15 +56,11 @@ export const crate = (config?: CrateConfig) => {
   const query = (sql: string, options?: QueryOptions) => {
     const { args, coltypes = false } = options || {}
     debug('query - sql %s args %O', sql, args)
-    return retry(
-      () =>
-        fetch(maybeShowColTypes(sqlEndpoint, coltypes), {
-          method: 'post',
-          body: JSON.stringify({ stmt: sql, ...(args && { args }) }),
-          headers: { 'Content-Type': 'application/json', ...authHeader },
-        }).then((res) => res.json() as Promise<QueryResult | ErrorResult>),
-      retryOptions
-    )
+    return fetch(maybeShowColTypes(sqlEndpoint, coltypes), {
+      method: 'post',
+      body: JSON.stringify({ stmt: sql, ...(args && { args }) }),
+      headers: { 'Content-Type': 'application/json', ...authHeader },
+    }).then((res) => res.json() as Promise<QueryResult | ErrorResult>)
   }
 
   const insert = (qualifiedName: string, record: object) => {
@@ -99,15 +81,11 @@ export const crate = (config?: CrateConfig) => {
   const bulkInsert = (qualifiedName: string, records: object[]) => {
     const { sql, args } = getBulkInsertSqlAndArgs(qualifiedName, records)
     debug('bulkInsert - sql %s bulk_args %O', sql, args)
-    return retry(
-      () =>
-        fetch(sqlEndpoint, {
-          method: 'post',
-          body: JSON.stringify({ stmt: sql, bulk_args: args }),
-          headers: { 'Content-Type': 'application/json', ...authHeader },
-        }).then((res) => res.json() as Promise<BulkQueryResult | ErrorResult>),
-      retryOptions
-    )
+    return fetch(sqlEndpoint, {
+      method: 'post',
+      body: JSON.stringify({ stmt: sql, bulk_args: args }),
+      headers: { 'Content-Type': 'application/json', ...authHeader },
+    }).then((res) => res.json() as Promise<BulkQueryResult | ErrorResult>)
   }
 
   return { query, insert, upsert, bulkInsert, deleteById }